[Fix] make some lib functions overloadable

+++ lib/memcmp.cpp
@@ -1,6 +1,6 @@
 #include <fried/string.h>
 
-int __attribute__((weak)) memcmp(const void* aptr, const void* bptr, size_t size) {
+int __attribute__((weak)) memcmp(const void *aptr, const void *bptr, unsigned long long size) {
 	const unsigned char* a = (const unsigned char*) aptr;
 	const unsigned char* b = (const unsigned char*) bptr;
 	for (size_t i = 0; i < size; i++) {
